// src/models/crypto/CryptoData.ts
export interface DashboardData {
  id: number;
  symbol: string;
  marketType: string;
  imgurl: string;
  active: boolean;
}

export interface CryptoData {
  dashboardData: DashboardData; // Dados aninhados
  priceChange_1m: string;
  priceChange_5m: string;
  priceChange_15m: string;
  priceChange_1h: string;
  priceChange_1d: string;
  pullbackEntry_5m: string | null;
  pullbackEntry_15m: string | null;
  pullbackEntry_1h: string | null;
  pullbackEntry_4h: string | null;
  pullbackEntry_1d: string | null;
  superTrend4h_5m: string | null;
  superTrend4h_15m: string | null;
  superTrend4h_1h: string | null;
  trueRange_1m: string;
  trueRange_5m: string;
  trueRange_15m: string;
  trueRange_1h: string;
  rsi_5m: string;
  rsi_15m: string;
  rsi_1h: string;
  rsi_4h: string;
  rsi_1d: string;
  ematrend_15m: string;
  ematrend_1h: string;
  ematrend_4h: string;
  ematrend_1d: string;
}